refactor(FOROME-722): split select and edit modals into one
