--- src/core/enum/modal-type-enum.ts
@@ -1,4 +1,0 @@
-export enum ModalTypeEnum {
-  Edit = 'Edit',
-  Select = 'Select',
-}
