+++ src/pages/filter/page.tsx
@@ -21,8 +21,7 @@ import { ModalEditCompoundRequest } from './ui/modal-edit/components/modal-edit-
 import { ModalEditCustomInheritanceMode } from './ui/modal-edit/components/modal-edit-custom-inheritance-mode'
 import { ModalEditFilters } from './ui/modal-edit/components/modal-edit-filters'
 import { ModalEditGeneRegion } from './ui/modal-edit/components/modal-edit-gene-region'
-import { ModalEditInheritanceMode } from './ui/modal-edit/components/modal-inheritance-mode/modal-edit-inheritance-mode'
-import { ModalSelectInheritanceMode } from './ui/modal-edit/components/modal-inheritance-mode/modal-select-inheritance-mode'
+import { ModalInheritanceMode } from './ui/modal-edit/components/modal-inheritance-mode/modal-inheritance-mode'
 import { ModalTextEditor } from './ui/query-builder/modal-text-editor'
 import { QueryBuilder } from './ui/query-builder/query-builder'
 import { ModalSaveDataset } from './ui/query-builder/ui/modal-save-dataset'
@@ -112,12 +111,7 @@ const FilterPage = observer((): ReactElement => {
 
       {dtreeStore.isModalNumbersVisible && <ModalNumbers />}
 
-      {dtreeStore.isModalEditInheritanceModeVisible && (
-        <ModalEditInheritanceMode />
-      )}
-      {dtreeStore.isModalSelectInheritanceModeVisible && (
-        <ModalSelectInheritanceMode />
-      )}
+      {dtreeStore.isModalInheritanceModeVisible && <ModalInheritanceMode />}
 
       {dtreeStore.isModalEditCustomInheritanceModeVisible && (
         <ModalEditCustomInheritanceMode />
