+++ src/pages/filter/ui/modal-edit/components/modal-inheritance-mode/modal-inheritance-mode.tsx
@@ -1,15 +1,17 @@
 import { ReactElement, useEffect, useRef, useState } from 'react'
 import { observer } from 'mobx-react-lite'
 
+import { ActionType } from '@declarations'
 import dtreeStore from '@store/dtree'
 import { InheritanceModeContent } from '@pages/filter/ui/modal-edit/components/modal-inheritance-mode/components/inheritance-mode-content'
 import { HeaderModal } from '@pages/filter/ui/query-builder/ui/header-modal'
 import { ModalBase } from '@pages/filter/ui/query-builder/ui/modal-base'
+import { SelectModalButtons } from '@pages/filter/ui/query-builder/ui/select-modal-buttons'
 import modalEditStore from '../../modal-edit.store'
 import { EditModalButtons } from '../edit-modal-buttons'
 import modalInheritanceModeStore from './modal-inheritance-mode.store'
 
-export const ModalEditInheritanceMode = observer((): ReactElement => {
+export const ModalInheritanceMode = observer((): ReactElement => {
   const ref = useRef(null)
 
   const {
@@ -18,23 +20,26 @@ export const ModalEditInheritanceMode = observer((): ReactElement => {
     currentGroup,
     groupName,
     selectedGroupsAmount,
-    currentGroupLength,
     problemGroups,
   } = modalEditStore
 
   const [selectedProblemGroups, setSelectedProblemGroups] = useState<string[]>(
-    (Object.values(currentGroup[currentGroup.length - 1])[0] as string[]) || [],
+    currentGroup
+      ? (Object.values(currentGroup[currentGroup.length - 1])[0] as string[])
+      : [],
   )
 
   useEffect(() => {
-    dtreeStore.stepData[currentStepIndex].groups[currentGroupIndex]
-      .find((elem: any) => Array.isArray(elem))
-      .map((item: string) => dtreeStore.addSelectedFilter(item))
+    if (currentGroup) {
+      dtreeStore.stepData[currentStepIndex].groups[currentGroupIndex]
+        .find((elem: any) => Array.isArray(elem))
+        .map((item: string) => dtreeStore.addSelectedFilter(item))
+    }
 
     return () => {
       dtreeStore.resetSelectedFilters()
     }
-  }, [currentGroupIndex, currentGroupLength, currentStepIndex])
+  }, [currentGroup, currentGroupIndex, currentStepIndex])
 
   useEffect(() => {
     const params = `{"problem_group":["${selectedProblemGroups
@@ -60,6 +65,10 @@ export const ModalEditInheritanceMode = observer((): ReactElement => {
     )
   }
 
+  const addAttribute = (action: ActionType) => {
+    modalInheritanceModeStore.addAttribute(action, selectedProblemGroups)
+  }
+
   return (
     <ModalBase refer={ref} minHeight={340}>
       <HeaderModal
@@ -79,16 +88,27 @@ export const ModalEditInheritanceMode = observer((): ReactElement => {
         }
       />
 
-      <EditModalButtons
-        handleClose={() => modalInheritanceModeStore.closeModal()}
-        handleSaveChanges={() =>
-          modalInheritanceModeStore.saveChanges(selectedProblemGroups)
-        }
-        disabled={
-          selectedGroupsAmount.length === 0 ||
-          selectedProblemGroups.length === 0
-        }
-      />
+      {currentGroup ? (
+        <EditModalButtons
+          handleClose={() => modalInheritanceModeStore.closeModal()}
+          handleSaveChanges={() =>
+            modalInheritanceModeStore.saveChanges(selectedProblemGroups)
+          }
+          disabled={
+            selectedGroupsAmount.length === 0 ||
+            selectedProblemGroups.length === 0
+          }
+        />
+      ) : (
+        <SelectModalButtons
+          handleClose={() => modalInheritanceModeStore.closeModal()}
+          handleModals={() => modalInheritanceModeStore.openModalAttribute()}
+          handleModalJoin={() => dtreeStore.openModalJoin()}
+          handleAddAttribute={addAttribute}
+          disabled={dtreeStore.selectedFilters.length === 0}
+          currentGroup={currentGroup}
+        />
+      )}
     </ModalBase>
   )
 })
